🔧 Change: StackList와 StackTag 컴포넌트 하나로 합침
